fix circular import and cropped labels
